Move public headers into new directory such that they can be included internally or via submdoule as: include dronecode_sdk.h include `plugins/<name>/<name>.h`
