+++ core/include/connection_result.h
