+++ core/include/dronecode_sdk.h
