+++ core/include/plugin_base.h
