+++ plugins/action/CMakeLists.txt
@@ -11,9 +11,16 @@ set_target_properties(dronecode_sdk_action
     PROPERTIES COMPILE_FLAGS ${warnings}
 )
 
+target_include_directories(dronecode_sdk_action 
+PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}
+)
+
 install(FILES
-    action.h
-    action_result.h
+    include/plugins/action/action.h
+    include/plugins/action/action_result.h
     DESTINATION ${dronecode_sdk_install_include_dir}
 )
