+++ plugins/calibration/calibration.cpp
@@ -1,4 +1,4 @@
-#include "calibration.h"
+#include "plugins/calibration/calibration.h"
 #include "calibration_impl.h"
 #include "calibration_messages.h"
