+++ plugins/camera/include/plugins/camera/camera.h
