+++ plugins/follow_me/CMakeLists.txt
@@ -12,7 +12,7 @@ set_target_properties(dronecode_sdk_follow_me
 )
 
 install(FILES
-    follow_me.h
+    include/plugins/follow_me/follow_me.h
     DESTINATION ${dronecode_sdk_install_include_dir}
 )
 
@@ -20,3 +20,11 @@ install(TARGETS dronecode_sdk_follow_me
     #EXPORT dronecode_sdk-targets
     DESTINATION ${dronecode_sdk_install_lib_dir}
 )
+
+target_include_directories(dronecode_sdk_follow_me
+PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}
+)
+
