+++ plugins/follow_me/include/plugins/follow_me/follow_me.h
