+++ plugins/gimbal/gimbal.cpp
@@ -1,4 +1,4 @@
-#include "gimbal.h"
+#include "plugins/gimbal/gimbal.h"
 #include "gimbal_impl.h"
 
 namespace dronecode_sdk {
