+++ plugins/gimbal/include/plugins/gimbal/gimbal.h
