+++ plugins/info/CMakeLists.txt
@@ -12,11 +12,18 @@ set_target_properties(dronecode_sdk_info
 )
 
 install(FILES
-    info.h
+    include/plugins/info/info.h
     DESTINATION ${dronecode_sdk_install_include_dir}
 )
 
 install(TARGETS dronecode_sdk_info
     #EXPORT dronecode_sdk-targets
     DESTINATION ${dronecode_sdk_install_lib_dir}
+)
+
+target_include_directories(dronecode_sdk_info
+PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}
 )
