+++ plugins/info/include/plugins/info/info.h
