+++ plugins/info/info_impl.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "info.h"
+#include "plugins/info/info.h"
 #include "mavlink_include.h"
 #include "plugin_impl_base.h"
 #include <mutex>
