+++ plugins/logging/logging.cpp
@@ -1,4 +1,4 @@
-#include "logging.h"
+#include "plugins/logging/logging.h"
 #include "logging_impl.h"
 
 namespace dronecode_sdk {
