+++ plugins/logging/logging_impl.h
@@ -3,7 +3,7 @@
 #include "plugin_impl_base.h"
 #include "mavlink_include.h"
 #include "system.h"
-#include "logging.h"
+#include "plugins/logging/logging.h"
 
 namespace dronecode_sdk {
