+++ plugins/mission/mission_impl.h
@@ -6,7 +6,7 @@
 
 #include "system.h"
 #include "mavlink_include.h"
-#include "mission.h"
+#include "plugins/mission/mission.h"
 #include "plugin_impl_base.h"
 #include <json11.hpp>
