+++ plugins/mission/mission_import_qgc_test.cpp
@@ -6,7 +6,7 @@
 #include <gtest/gtest.h>
 #include <atomic>
 #include <cmath>
-#include "mission.h"
+#include "plugins/mission/mission.h"
 #include "mavlink_include.h"
 #include "global_include.h"
 #include "log.h"
