+++ plugins/mission/mission_item.cpp
@@ -1,4 +1,4 @@
-#include "mission_item.h"
+#include "plugins/mission/mission_item.h"
 
 #include <iomanip>
 #include <iostream>
