+++ plugins/telemetry/CMakeLists.txt
@@ -13,11 +13,18 @@ set_target_properties(dronecode_sdk_telemetry
 )
 
 install(FILES
-    telemetry.h
+    include/plugins/telemetry/telemetry.h
     DESTINATION ${dronecode_sdk_install_include_dir}
 )
 
 install(TARGETS dronecode_sdk_telemetry
     #EXPORT dronecode_sdk-targets
     DESTINATION ${dronecode_sdk_install_lib_dir}
+)
+
+target_include_directories(dronecode_sdk_telemetry
+PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}
 )
