+++ plugins/telemetry/telemetry.cpp
@@ -1,4 +1,4 @@
-#include "telemetry.h"
+#include "plugins/telemetry/telemetry.h"
 
 #include <limits>
 #include <cmath>
