+++ plugins/telemetry/telemetry_impl.h
@@ -2,7 +2,7 @@
 
 #include <atomic>
 #include <mutex>
-#include "telemetry.h"
+#include "plugins/telemetry/telemetry.h"
 #include "plugin_impl_base.h"
 #include "system.h"
 #include "mavlink_include.h"
